Deprecated `mrsa_source_bacteraemia`
`mrsa_source_bacteraemia` was too specific to MRSA and had been
generalised to `group_source_bacteraemia` when new Gram negative
collections were introduced.
As a result `mrsa_source_bacteraemia` is deprecated.

`group_source_bacteraemia` was altered so that NA is recoded to
"Not reported".

Also added minor updates to readme to add links to github and code of
conduct.